requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
et_xmlfile==2.0.0
idna==3.11
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
pillow==12.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
reportlab==4.4.10
requests==2.33.1
six==1.17.0
tqdm==4.67.3
tzdata==2026.1
urllib3==2.6.3
